python-backend/requirements.txt
websockets==12.0
numpy==1.26.2
torch==2.1.0
openai-whisper